requirements.txt
"""
Requirements for the RDR2 Agent API production deployment.
Install with: pip install -r requirements.txt
"""

# Core dependencies from your existing system
crewai>=0.80.0
google-generativeai>=0.8.0
chromadb>=0.5.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
agentops>=0.3.0

# FastAPI and web server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Production server dependencies
gunicorn>=21.2.0
prometheus-client>=0.19.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
requests>=2.31.0

# Monitoring and logging
structlog>=23.2.0
python-json-logger>=2.0.0

# Security
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Optional: Rate limiting
slowapi>=0.1.9
redis>=5.0.0
